First Test to Get Default Statistics

+++ NimatorCouchBase/Entities/Statistics/Ram.cs
@@ -4,10 +4,10 @@ namespace NimatorCouchBase.Entities.Statistics
     {
         public long Total { get; set; }
         public long QuotaTotal { get; set; }
-        public int QuotaUsed { get; set; }
+        public long QuotaUsed { get; set; }
         public long Used { get; set; }
-        public int UsedByData { get; set; }
-        public int QuotaUsedPerNode { get; set; }
+        public long UsedByData { get; set; }
+        public long QuotaUsedPerNode { get; set; }
         public long QuotaTotalPerNode { get; set; }
     }
 }
